Also register constructors as functions on the class namespace
There can be multiple constructors
(typically of form "my_object_new_from_X"). These where previously
not available in JavaScript bindings.
The normal constructor should also be available
as "Something.new()" in addition to as "new Something()".
This matches the behavior of for instance PyGObject. This currently
has a bug, see the tests added in previous commit.